// customer-service/src/utils/enums/error-code.ts
export enum ErrorCode {
    CUSTOMER_NOT_FOUND = "CUSTOMER_NOT_FOUND",
    SENDER_NOT_FOUND = "SENDER_NOT_FOUND",
    RECEIVER_NOT_FOUND = "RECEIVER_NOT_FOUND",

    INVALID_DEPOSIT_AMOUNT = "INVALID_DEPOSIT_AMOUNT",
    INVALID_PURCHASE_AMOUNT = "INVALID_PURCHASE_AMOUNT",
    INSUFFICIENT_FUNDS = "INSUFFICIENT_FUNDS",

    INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
}